feat(sidebar): bring back the narrow sidebar as a Slack-style rail
Adds a `narrow` sidebar state: a 72px rail of big buttons — a wide icon
tile with the feature name underneath — one per feature, in place of the
wide list. Search leads the rail, then Home / Getting Started / Inbox /
Activity, then the workspace links the user kept; hidden workspace links
and the "Try" shortcuts move into a More menu, and Create plus the
account tile pin to the bottom.

The wide sidebar's header gains a toggle into the rail, and the rail's
logo tile toggles back. The choice is persisted, so hiding the sidebar
(cmd+.) and showing it again restores whichever variant was in use.
Anything that needs full width — favorites, recent channels, promo
cards, per-inbox rows, the in-call panel — stays with the wide variant;
on the rail the call UI floats over the app, as it already does when
the sidebar is collapsed.

Link data, the navigation helper, the go-to hotkey signal and the
persisted sidebar customization move to `sidebar-links.ts`, and the
dropdown link and settings widget to their own modules, so both
variants share them without a circular import.
